Make the guess input take more horizontal space
I deleted the submit button from this form, because I read that the
default behaviour for a form with one input is to submit on enter. I
also hid the label and added a placeholder. I'm not 100% certain that
this move is the right one in terms of people being able to discover how
to use it, but I prefer this aesthetically so I'm going with it for now.