Move focus to success message on next tick of the event loop
Something about focusing it (causing the keyboard to collapse on my
phone) and inserting it at the same time causes an annoying problem
where the message will appear very briefly at the top of the screen and
jump down. If I don't remove focus from the input bar, the text doesn't
jump, so I think if I delay moving focus, it will also fix the problem.

It might be more correct to move focus to the nextWord button instead,
since the experience will be the same on desktop, and won't matter on
mobile since you can just tap the success message. I kind of wanted to
delete the buttons because I'm lazy and don't want to figure out how to
style them to look good, but I think their removal might make the app
too restrictive (unrelated to the page jump problem.)