Use a symbol for guess input submit button
This is kind of a dumb reason: When the button has more text in it, part
of it does not stay inside the main container, which makes the other
content look smaller. I'm not sure why the input doesn't size itself so
as to fit inside the container.

The current code still has that problem, but the size of screen that it
happens on is now smaller than my phone width, so it works on my machine
for now.