Update colors so there's not quite as much black
Also move some shared styles to a shared place.

I'm not sure what to do with that big space in between the retry/next
buttons. Was considering putting a slider in there for the speed, so
everything is above the fold, but I suspect that on a phone that's not
enough room. We'll see.