Set up eslint
Kind of annoyed that this requires a package.json even if you want to
use eslint as a global.

I was unsure what to do about the event handlers where I didn't use the
event, but decided that it wasn't that weird to remove the argument in
that case. Still feel a little conflicted about it though -- sort of
wish there was an option for "my function is essentially implementing an
interface, so don't complain about the event arg".